Fix build: use int64_t strides in allocateUninitialized
main uplifted tt-mlir so createOwnedHostTensor's stride argument is now
std::vector<int64_t> (shape stays uint32_t), and calculateStrides returns
int64_t accordingly. The newly-added allocateUninitialized path still
declared its strides vector as uint32_t, which failed to compile against
the uplifted runtime API after merging main.

+++ pjrt_implementation/src/api/buffer_instance.cc
@@ -376,7 +376,7 @@ void BufferInstance::allocateUninitialized(const std::int64_t *byte_strides,
       tt::runtime::utils::dataTypeElementSize(runtime_data_type);
   std::vector<std::uint32_t> shape =
       calculateShape(getDimensionsRaw(), getNumberOfDimensions(), m_data_type);
-  std::vector<std::uint32_t> strides = calculateStrides(
+  std::vector<std::int64_t> strides = calculateStrides(
       getNumberOfDimensions(), byte_strides, num_byte_strides, element_size);
 
   // Allocate an owned host tensor without copying any data in. The contents are
